Move ObjectNames to common module.
