+++ java/sidecar/src/main/java/com/instaclustr/sidecar/cassandra/cassandra/CassandraObjectNames.java
@@ -1,5 +1,7 @@
 package com.instaclustr.sidecar.cassandra.cassandra;
 
+import com.instaclustr.jmx.ObjectNames;
+
 import javax.management.ObjectName;
 
 public final class CassandraObjectNames {
